+++ src/watson/analysis/CoreProtectAnalysis.java
@@ -75,6 +75,7 @@ public class CoreProtectAnalysis extends Analysis
       _x = Integer.parseInt(m.group(1));
       _y = Integer.parseInt(m.group(2));
       _z = Integer.parseInt(m.group(3));
+      Controller.instance.selectPosition(_x, _y, _z);
       _firstInspectorResult = true;
     }
   } // inspectorCoords