+++ src/watson/macro/ScriptActionWatson.java
@@ -1,0 +1,63 @@
+package watson.macro;
+
+import net.eq2online.macros.scripting.ScriptAction;
+import net.eq2online.macros.scripting.ScriptCore;
+import net.eq2online.macros.scripting.api.APIVersion;
+import net.eq2online.macros.scripting.api.IMacro;
+import net.eq2online.macros.scripting.api.IMacroAction;
+import net.eq2online.macros.scripting.api.IScriptActionProvider;
+import net.minecraft.src.mod_ClientCommands;
+
+// ----------------------------------------------------------------------------
+/**
+ * A Macro/Keybind Mod script action for a WATSON(&lt;string&gt;) command.
+ */
+@APIVersion(9)
+public class ScriptActionWatson extends ScriptAction
+{
+  // --------------------------------------------------------------------------
+  /**
+   * This method is called by reflection when Watson is loaded in order to
+   * attempt to register custom script actions.
+   * 
+   * @throws NoClassFoundDefError if the Macro/Keybind Mod is not loaded.
+   */
+  public static void initialise()
+  {
+    new ScriptActionWatson().OnInit();
+  }
+
+  // --------------------------------------------------------------------------
+  /**
+   * Default constructor.
+   */
+  public ScriptActionWatson()
+  {
+    super("watson");
+  }
+
+  // --------------------------------------------------------------------------
+  /**
+   * @see net.eq2online.macros.scripting.api.IMacrosAPIModule#OnInit()
+   */
+  @Override
+  public void OnInit()
+  {
+    ScriptCore.RegisterScriptAction(this);
+  }
+
+  // --------------------------------------------------------------------------
+  /**
+   * @see net.eq2online.macros.scripting.api.IScriptAction#Execute(net.eq2online.macros.scripting.api.IScriptActionProvider,
+   *      net.eq2online.macros.scripting.api.IMacro,
+   *      net.eq2online.macros.scripting.api.IMacroAction, java.lang.String,
+   *      java.lang.String[])
+   */
+  @Override
+  public String Execute(IScriptActionProvider provider, IMacro macro,
+                        IMacroAction action, String command, String[] args)
+  {
+    mod_ClientCommands.getInstance().handleClientCommand(command);
+    return null;
+  }
+} // class ScriptSctionWatsonCommand